chore: Move getContracts to Refinable class
